requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langgraph
python-dotenv
fastapi
uvicorn
mysql-connector-python
bcrypt
python-jose
passlib
pinecone
langchain_pinecone
redis